move get_entity* helper function to test
